use #+SETUPFILE instead of #+INCLUDE
Quoted from the Org Manual: #+INCLUDE only takes into effect during export. #+SETUPFILE defines a file that holds more in-buffer setup. Normally this is entirely ignored. Only when the buffer is parsed for option-setting lines (i.e., when starting Org mode for a file, when pressing C-c C-c in a settings line, or when exporting), then the contents of this file are parsed as if they had been included in the buffer. In particular, the file can be any other Org mode file with internal setup. You can visit the file the cursor is in the line with C-c '.